[REFACTOR] Statement 내부의 중첩클래스 top level 클래스로 모두 이동 및 접근제한자 package-private으로 변경
- PerformanceCalculator
- PerformanceCalculatorFactory
- TragedyCalculator
- ComedyCalculator
