+++ src/main/java/com/tkl/refactoring/chapter1/statement/domain/StatementPerformance.java
@@ -32,70 +32,4 @@ public class StatementPerformance {
 	public int volumeCredits() {
 		return volumeCredits;
 	}
-
-	static abstract class PerformanceCalculator {
-		protected final int audience;
-
-		public PerformanceCalculator(int audience) {
-			this.audience = audience;
-		}
-
-		abstract int amountFor();
-
-		final int volumeCreditsFor() {
-			return Math.max(this.audience - 30, 0) + extraVolumeCreditsFor();
-		}
-
-		protected int extraVolumeCreditsFor() {
-			return 0;
-		}
-	}
-
-	static class TragedyCalculator extends PerformanceCalculator {
-
-		public TragedyCalculator(int audience) {
-			super(audience);
-		}
-
-		@Override
-		public int amountFor() {
-			int result = 40000;
-			if (super.audience > 30) {
-				result += 1000 * (super.audience - 30);
-			}
-			return result;
-		}
-	}
-
-	static class ComedyCalculator extends PerformanceCalculator {
-
-		public ComedyCalculator(int audience) {
-			super(audience);
-		}
-
-		@Override
-		public int amountFor() {
-			int result = 30000;
-			if (super.audience > 20) {
-				result += 10000 + 500 * (super.audience - 20);
-			}
-			result += 300 * super.audience;
-			return result;
-		}
-
-		@Override
-		protected int extraVolumeCreditsFor() {
-			return this.audience / 5;
-		}
-	}
-
-	private static class PerformanceCalculatorFactory {
-		public static PerformanceCalculator create(int audience, Play play) {
-			return switch (play.type()) {
-				case "tragedy" -> new TragedyCalculator(audience);
-				case "comedy" -> new ComedyCalculator(audience);
-				default -> throw new Error("unknown type: ${play.type}");
-			};
-		}
-	}
 }
